Creada clase para output visual y sus tests. Varios fix
