+++ src/main/java/com/dir/inno/normalizacion_pdfs/Main.java
@@ -13,6 +13,7 @@ import org.xml.sax.SAXException;
 import com.snowtide.PDF;
 import com.snowtide.pdf.Document;
 import com.snowtide.pdf.OutputTarget;
+import com.snowtide.pdf.VisualOutputTarget;
 import java.io.File;
 import java.text.ParseException;
 import pdfts.examples.XMLOutputTarget;
@@ -34,7 +35,7 @@ public class Main {
     public static void main(String[] args) throws IOException, TransformerException, ParserConfigurationException, SAXException, ParseException {
 
         String filePath = new File("").getAbsolutePath();
-        filePath += "\\Almacenamiento PDFs editables\\PDFs editables\\Formulario de presentacion v4.7(5)_impreso.pdf";
+        filePath += "\\Almacenamiento PDFs editables\\PDFs editables\\Formulario de presentacion v4.7(1)_impreso.pdf";
         
         StringBuilder text;
         try (Document pdf = PDF.open(filePath)) {
@@ -45,15 +46,19 @@ public class Main {
         }
 
         LectorPDFImpreso47 lector = new LectorPDFImpreso47(text);
-
-        /*
-        Transformer transformer = TransformerFactory.newInstance().newTransformer();
-        Source input = new DOMSource(xml.getXMLDocument());
-        Result output = new StreamResult(new File("output.xml"));
-        transformer.transform(input, output);
-         */
+        
+        try (Document pdf = PDF.open(filePath)) {
+            XMLOutputTarget xml = new XMLOutputTarget();
+            pdf.pipe(xml);
+            text = new StringBuilder();
+            pdf.pipe(new VisualOutputTarget(text));
+        }
+        
+        //MOSTRAR SALIDA
         System.out.println(text);
-
+        
+        LectorPDFImpreso47V lectorV = new LectorPDFImpreso47V(text);
+        
         System.out.println("Version: " + lector.obtenerVersion() + "\r");
 
         System.out.println("Nombre: " + lector.obtenerNombre());
@@ -71,7 +76,7 @@ public class Main {
 
         System.out.println("\rDomicilio Constituido: " + lector.obtenerDomicilioConst() + "\r");
 
-        ArrayList<ArrayList<String>> nomina = lector.obtenerNomina();
+        ArrayList<ArrayList<String>> nomina = lectorV.obtenerNomina();
         for (Integer i = 0; i < nomina.size(); i++) {
             System.out.println("Autoridad Societaria " + (i + 1) + '/' + nomina.size() + ':');
             System.out.println(nomina.get(i).get(0) + ' '
@@ -99,7 +104,7 @@ public class Main {
         System.out.println("\nNombre archivo foto satelital: " + lector.obtenerNombreArchivoFotoSat());
 
         
-        ArrayList<ArrayList<String>> partidas = lector.obtenerPartidasInm();
+        ArrayList<ArrayList<String>> partidas = lectorV.obtenerPartidasInm();
         for (Integer i = 0; i < partidas.size(); i++) {
             System.out.println("Partida inmobiliaria " + (i + 1) + '/'
                     + partidas.size() + ": " + partidas.get(i).get(0));
